requirements.txt
Django
psycopg2
djangorestframework
pygments
